Modify "signals" to be a unified "port"
This fixes and simplifies alot of things;
Ports can be connected to other ports or to a propagation function. Removes the need for separate input/output signals, (signals are now implicit in the linked ports available for a port).
Furthermore, this presents a new clocking algorithm, wherein values now propagate outwards from registers, through components. Previously, Components were the starting point, wherein the registers were searched for, through its inputs. There is still overhead and room for improvement in this algorithm, but it works for now.