Make components non-templated
This change address two things:
- Ease of interfacing into core components from graphics components
- Allow parsing files at runtime, and create core components.

The obvious drawback is the loss of compile-time checks when connecting components.
However, given that this library is intended to be able to at runtime parse a HDL file and create the resulting graphical simulator, we cannot resort to templated base components, unless an intermediate compiler/parser will parse HDL files into C++ files with VSRTL syntax.
This is not desired, given that this increases the development burden as well as makes the program harder to use. This might also result in a performance hit, but again, the intention is not to create a highly performant simulator.