Introduce vsrtl::Base class
This class shall be used to create a registration method for the graphics library to register the corresponding graphical component for a component, with the component itself - without creating a dependency between the Graphics library and the Core library.

This allows for separation of concerns between the graphics and the Core library - by being able to go from Core component to Graphic component, all graph traversal shall be done in the Core domain, and if required, the graphical components can be accesed through the methods present in the vsrtl::Base class.