Make source/sink graphic ports share state #21
This commit introduced a shared selection state between a string of source/sink ports, wherein it is the origin source port which maintains the state of the string of connected ports and wires.
In this, the origin source port will maintain the pen which the string of connected ports is drawn with, as well as selection state.