Add DisplayType actions to TreeItem's
TreeItem's can now be queried for their actions, which may be presented
when requesting a menu in related to a TreeItem.
TreeItem presents DisplayType actions.
Subclasses may override the getActions() function to extend the number of
provided actions, as well as call TreeItem::getActions() to make the default
actions available.