Implement propagation stack #10
This commit introduces the propagation stack, a stack of Port* which are
created by the design propagation algorithm.
This stack contains the sequence of which ports shall be updated, for input
dependencies of ports to be resolved. Thus, the stack must be traversed
sequentially.

Future improvements may come in partitioning parts of this stack, and executing
the propagation logic for "strings" of independent ports and components on
separate threads.