[WIP] Move grid logic out of graphics components
This commit extracts the grid-related logic from ComponentGraphic and
PortGraphic into separate objects within the EDA library.
The end goal of this being to have the graphics objects purely be a reflection
of the state of the grid components.
The grid components are what the EDA algorithms will work on, and be
represened in grid coordinates, as opposed to scene coordinates (graphics)
which may be arbitrarily scaled.