Make Port widths a template parameter
This significantly improves static-error checking when connecting ports, by
leveraging the C++ type system.

However, somewhat increases complexity in terms of interacting with the graphical
library.
For components which requires custom graphics (ALU, Registers), these has to
be defined as non-templated base classes, to ensure that the graphical component
is equal for any type of templated subclass.