Remove assert on Memory force value
The assert caught a non-issue; forcing a value is a modification of
the current state of the memory component. The current state has yet
to be pushed onto the rewind stack, and as such does not modify
the semantics of the circuit if a rewind is performed after a value
has been forced onto the memory component.