Rework source-sink port visibility logic
Now, when the source port of a given port is non-visible, sink ports will still
be visible in the scene, but will not be drawn.
From this, we may still have labels and indicators associated with a given sink
port be visible, even though the port is not being drawn.
Furthermore, since the source port is drawn when the user hovers over the port,
it is possible to move the port (and hence move the label and indicator aswell)
whilst its source is non visible.