Initial commit for dynamically instantiated circuits
Dynamic circuits are able to be specified at runtime.
To implement this, Ports are no longer templated with a port width,
but instead receive a width value at runtime. Doing this, we trade
arguably nice behaviour of detecting port mismatches at compile time,
for the ability to dynamically instantiate ports.

Dynamic components (lets call them 'operators') will be instantiated in
the 'ops' folder. A few base operators are provided along with this
commit.

An example dynamically specified circuit is shown in the "app.cpp" file.