Translate Gallant signal/slots through Qt signal slot system
Given that the simulator signal/slot framework is not the Qt framework, we here
provide a translation signal for simulator-to-Qt signals. This is done to allow
for using Qt's Qt::AutoConnection connection type which handles cross-thread
signal-slot connections if the simulator is executing in some none-GUI thread.
Handling cross-thread signals in turns fixed a bug in which Qt objects which
must be created on the GUI thread are inadvertently created on a non-gui
thread through handling the simulator signal.