[WIP] Make value labels of ports drawn as top-level within parent component
Previously, value labels of ports could be nested behind wires
or other components arbitrarily, based on the sequence of
construction of objects.
This commit moves ownership of ports' value labels to the
enclosing component.
This brings a variety of challenges, because we still want the
port itself to "act" like the parent component of the label
(port hidden/moved => hide/move label etc.). To handle this
we introduce the concept of virtual parent/child relationship.