Implement virtualChild QGraphicsItems for mirroring item movements
This is needed to have better control over the layering of different items. For instance, value labels should be drawn on top of all wires and components, so they need to be a child of the enclosing module. However, these value labels should still move with the port which they are associated with, as well as the component which owns the port.